rag_ops/requirements.txt
# Existing requirements
streamlit
langchain
langchain-community
langchain-openai
python-dotenv
PyMuPDF
pandas
aiofiles
asyncio
chromadb

fastapi
uvicorn
python-multipart
httpx
pydantic-settings
python-multipart

# Testing dependencies
pytest==7.4.4
pytest-asyncio==0.23.5
httpx==0.26.0
python-multipart==0.0.7
aiofiles==23.2.1


